+++ src/com/maddyhome/idea/vim/action/motion/updown/MotionArrowUpAction.kt
@@ -28,7 +28,10 @@ import com.maddyhome.idea.vim.command.Command
 import com.maddyhome.idea.vim.command.CommandFlags
 import com.maddyhome.idea.vim.command.MappingMode
 import com.maddyhome.idea.vim.handler.NonShiftedSpecialKeyHandler
+import com.maddyhome.idea.vim.helper.EditorHelper
 import com.maddyhome.idea.vim.helper.StringHelper.parseKeys
+import com.maddyhome.idea.vim.helper.isEndAllowed
+import com.maddyhome.idea.vim.helper.mode
 import com.maddyhome.idea.vim.helper.vimLastColumn
 import java.awt.event.KeyEvent
 import java.util.*
@@ -47,7 +50,9 @@ private object MotionArrowUpActionHandler : NonShiftedSpecialKeyHandler() {
   }
 
   override fun postMove(editor: Editor, caret: Caret, context: DataContext, cmd: Command) {
-    caret.vimLastColumn = col
+    val pos = caret.visualPosition
+    val lastColumn = EditorHelper.lastColumnForLine(editor, pos.line, editor.mode.isEndAllowed)
+    caret.vimLastColumn = if (pos.column != lastColumn) pos.column else col
   }
 }