+++ src/com/maddyhome/idea/vim/action/motion/updown/MotionShiftUpAction.kt
@@ -26,6 +26,9 @@ import com.maddyhome.idea.vim.command.Command
 import com.maddyhome.idea.vim.command.MappingMode
 import com.maddyhome.idea.vim.group.MotionGroup
 import com.maddyhome.idea.vim.handler.ShiftedArrowKeyHandler
+import com.maddyhome.idea.vim.helper.EditorHelper
+import com.maddyhome.idea.vim.helper.isEndAllowed
+import com.maddyhome.idea.vim.helper.mode
 import com.maddyhome.idea.vim.helper.vimForEachCaret
 import com.maddyhome.idea.vim.helper.vimLastColumn
 import javax.swing.KeyStroke
@@ -40,7 +43,11 @@ private object MotionShiftUpActionHandler : ShiftedArrowKeyHandler() {
       val vertical = VimPlugin.getMotion().moveCaretVertical(editor, caret, -cmd.count)
       val col = caret.vimLastColumn
       MotionGroup.moveCaret(editor, caret, vertical)
-      caret.vimLastColumn = col
+
+      val pos = caret.visualPosition
+      val lastColumn = EditorHelper.lastColumnForLine(editor, pos.line, editor.mode.isEndAllowed)
+      val targetColumn = if (pos.column != lastColumn) pos.column else col
+      caret.vimLastColumn = targetColumn
     }
   }