+++ src/com/maddyhome/idea/vim/helper/EditorHelper.java
@@ -292,7 +292,9 @@ public class EditorHelper {
       return getFileSize(editor, allowEnd);
     }
     else {
-      return editor.getDocument().getLineEndOffset(line) - (allowEnd ? 0 : 1);
+      final int startOffset = editor.getDocument().getLineStartOffset(line);
+      final int endOffset = editor.getDocument().getLineEndOffset(line);
+      return endOffset - (startOffset == endOffset || allowEnd ? 0 : 1);
     }
   }
 
@@ -711,6 +713,10 @@ public class EditorHelper {
       return scrollFullPageUp(editor, pages);
     }
     return -1;  // visual lines are 1-based
+  }
+
+  public static int lastColumnForLine(@NotNull final Editor editor, int line, boolean allowEnd) {
+    return editor.offsetToVisualPosition(EditorHelper.getLineEndOffset(editor, line, allowEnd)).column;
   }
 
   private static int scrollFullPageDown(@NotNull final Editor editor, int pages) {